Create PyRansomV1
A very basic ransomware written in Python. This is intended for research purposes only and is not intended to cause any harm. It is important to note that the script only targets files in the existing directory of the ransomware and does not launch itself (a victim would need to willingly download and run the malware in order for it to cause any damage). It is highly recommended to run this malware in an isolated environment in order to minimize any possible damage during testing. It is also important to remember that creating, distributing, or using ransomware is illegal and can have serious consequences, therefore I take NO accountability for this. Again, DISCLAIMER: This is intended for EDUCATIONAL PURPOSES ONLY!!